Moved structs to separate files
