+++ src/audio/pulse/mod.rs
@@ -1,0 +1,3 @@
+mod pulse_wave_pattern_duty;
+
+pub use self::pulse_wave_pattern_duty::PulseWavePatternDuty;
