+++ src/memory/audio_registers.rs
@@ -1,76 +1,7 @@
+use crate::audio::pulse::PulseWavePatternDuty;
 use crate::audio::sweep::Sweep;
 use crate::audio::{VolumeEnvelopeDirection, WaveOutputLevel};
 use crate::{Byte, Word};
-
-#[derive(Clone)]
-pub enum PulseWavePatternDuty {
-    Percent125,
-    Percent25,
-    Percent50,
-    Percent75,
-}
-
-impl PulseWavePatternDuty {
-    pub fn calculate_amplitude(&self, position: Byte) -> f32 {
-        match self {
-            Self::Percent125 => {
-                if position == 7 {
-                    1.0
-                } else {
-                    0.0
-                }
-            }
-            Self::Percent25 => {
-                if position > 5 {
-                    1.0
-                } else {
-                    0.0
-                }
-            }
-            Self::Percent50 => {
-                if position > 3 {
-                    1.0
-                } else {
-                    0.0
-                }
-            }
-            Self::Percent75 => {
-                if position < 6 {
-                    1.0
-                } else {
-                    0.0
-                }
-            }
-        }
-    }
-
-    pub fn to_percent(&self) -> f32 {
-        match self {
-            Self::Percent125 => 0.125,
-            Self::Percent25 => 0.25,
-            Self::Percent50 => 0.50,
-            Self::Percent75 => 0.75,
-        }
-    }
-}
-
-impl Default for PulseWavePatternDuty {
-    fn default() -> Self {
-        Self::Percent50
-    }
-}
-
-impl From<Byte> for PulseWavePatternDuty {
-    fn from(wave_duty: Byte) -> Self {
-        match wave_duty {
-            0b00 => Self::Percent125,
-            0b01 => Self::Percent25,
-            0b10 => Self::Percent50,
-            0b11 => Self::Percent75,
-            _ => panic!("Invalid Wave Duty"),
-        }
-    }
-}
 
 #[readonly::make]
 pub struct AudioRegisters {
